sym_handler.py decomposed
sym_cleanup.py is singled out to hold sym listener cleanup logic.
